use size_t instead of int for iteration variable
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@3139 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/layout/GEMLayout.cpp
@@ -337,7 +337,7 @@ bool GEMLayout::run() {
         std::vector<std::set<node> > components;
         string err;
         ConnectedTest::computeConnectedComponents(graph, components);
-        for (int i = 0; i < components.size(); ++i) {
+        for (size_t i = 0; i < components.size(); ++i) {
             Graph * tmp = graph->inducedSubGraph(components[i]);
             tmp->computeProperty("GEM (Frick)", layoutResult, err, pluginProgress, dataSet);
             graph->delAllSubGraphs(tmp);
